refactor: organize tests better
