+++ tests/Application/app/Group.php
@@ -13,11 +13,13 @@ class Group extends Model
 {
     use SoftDeletes;
 
+    /** @phpstan-return HasMany<Account> */
     public function accounts(): HasMany
     {
         return $this->hasMany(Account::class);
     }
 
+    /** @phpstan-return HasMany<User> */
     public function users(): HasMany
     {
         return $this->hasMany(User::class);
