+++ tests/Features/Methods/Builder.php
@@ -9,42 +9,46 @@ use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use Illuminate\Support\Collection;
 use stdClass;
+use function PHPStan\dumpType;
 
+/**
+ * This class tests `EloquentBuilder::__call` method
+ *
+ * So every test should either begin with `User::query`,
+ * or need to use a `Builder` typed argument.
+ *
+ * Return type should always be checked with `@return` annotation.
+ */
 class Builder
 {
+    /** @phpstan-return EloquentBuilder<User> */
     public function testGroupBy(): EloquentBuilder
     {
         return User::query()->groupBy('foo', 'bar');
     }
 
-    public function testDynamicWhereAsString(): ?User
+    /** @phpstan-return EloquentBuilder<User> */
+    public function testDynamicWhereAsString(): EloquentBuilder
     {
-        return (new User())->whereEmail('bar')->first();
+        return User::query()->whereEmail('bar');
     }
 
-    public function testDynamicWhereMultiple(): ?User
+    /** @phpstan-return EloquentBuilder<User> */
+    public function testDynamicWhereMultiple(): EloquentBuilder
     {
-        return User::whereIdAndEmail(1, 'foo@example.com')->first();
+        return User::query()->whereIdAndEmail(1, 'foo@example.com');
     }
 
-    public function testDynamicWhereAsInt(): ?User
+    /** @phpstan-return EloquentBuilder<User> */
+    public function testDynamicWhereAsInt(): EloquentBuilder
     {
-        return (new User())->whereEmail(1)->first();
-    }
-
-    public function testGetQueryReturnsQueryBuilder(): QueryBuilder
-    {
-        return User::getQuery();
-    }
-
-    public function testToBaseReturnsQueryBuilder(): QueryBuilder
-    {
-        return User::toBase();
+        return User::query()->whereEmail(1);
     }
 
     public function testToBaseReturnsQueryBuilderAfterChain(): QueryBuilder
     {
-        return User::whereNull('name')
+        return User::query()
+            ->whereNull('name')
             ->orderBy('email')
             ->toBase();
     }
@@ -74,46 +78,44 @@ class Builder
 
     public function testOrderByToBaseWithQueryExpression(): ?QueryBuilder
     {
-        return User::whereNull('name')
+        return User::query()
+            ->whereNull('name')
             ->orderBy(\Illuminate\Support\Facades\DB::raw('name'))
             ->toBase();
     }
 
     public function testLatestToBaseWithQueryExpression(): ?QueryBuilder
     {
-        return User::whereNull('name')
+        return User::query()
+            ->whereNull('name')
             ->latest(\Illuminate\Support\Facades\DB::raw('created_at'))
             ->toBase();
     }
 
     public function testOldestToBaseWithQueryExpression(): ?QueryBuilder
     {
-        return User::whereNull('name')
+        return User::query()
+            ->whereNull('name')
             ->oldest(\Illuminate\Support\Facades\DB::raw('created_at'))
             ->toBase();
     }
 
     public function testPluckToBaseWithQueryExpression(): ?Collection
     {
-        return User::whereNull('name')
+        return User::query()
+            ->whereNull('name')
             ->pluck(\Illuminate\Support\Facades\DB::raw('created_at'))
             ->toBase();
     }
 
     public function testIncrementWithQueryExpression(): int
     {
-        /** @var User $user */
-        $user = new User;
-
-        return $user->increment(\Illuminate\Support\Facades\DB::raw('counter'));
+        return User::query()->increment(\Illuminate\Support\Facades\DB::raw('counter'));
     }
 
     public function testDecrementWithQueryExpression(): int
     {
-        /** @var User $user */
-        $user = new User;
-
-        return $user->decrement(\Illuminate\Support\Facades\DB::raw('counter'));
+        return User::query()->decrement(\Illuminate\Support\Facades\DB::raw('counter'));
     }
 
     /** @phpstan-return null|EloquentBuilder<User> */
@@ -156,5 +158,60 @@ class Builder
         User::query()->unless($foo, static function (EloquentBuilder $query) {
             $query->whereNull('name');
         });
+    }
+
+    public function testMacro(\Illuminate\Database\Eloquent\Builder $query): void
+    {
+        $query->macro('customMacro', function () {
+        });
+    }
+
+
+    public function testFirstOrFailWithChain(): User
+    {
+        return User::with('foo')
+            ->where('email', 'bar')
+            ->orWhere('name', 'baz')
+            ->firstOrFail();
+    }
+
+    public function testFirstWithChain(): ?User
+    {
+        return User::with('foo')
+            ->where('email', 'bar')
+            ->orWhere('name', 'baz')
+            ->first();
+    }
+
+    public function testFirstWhere(): ?User
+    {
+        return User::query()->firstWhere(['email' => 'foo@bar.com']);
+    }
+
+
+    public function testOrWhereWithQueryExpression(): ?User
+    {
+        return User::with('foo')
+            ->orWhere(\Illuminate\Support\Facades\DB::raw('name'), 'like', '%john%')
+            ->first();
+    }
+
+    public function testWhereWithQueryExpression(): ?User
+    {
+        return User::with('foo')
+            ->where(\Illuminate\Support\Facades\DB::raw('name'), 'like', '%john%')
+            ->first();
+    }
+
+    public function testFirstWhereWithQueryExpression(): ?User
+    {
+        return User::with('foo')
+            ->firstWhere(\Illuminate\Support\Facades\DB::raw('name'), 'like', '%john%');
+    }
+
+    public function testValueWithQueryExpression(): ?string
+    {
+        return User::with('foo')
+            ->value(\Illuminate\Support\Facades\DB::raw('name'));
     }
 }
