+++ tests/Features/Models/Scopes.php
@@ -15,21 +15,25 @@ class Scopes extends Model
     /** @var User */
     private $user;
 
+    /** @phpstan-return HasOne<User> */
     public function testScopeAfterRelation(): HasOne
     {
         return $this->hasOne(User::class)->active();
     }
 
+    /** @phpstan-return HasMany<User> */
     public function testScopeAfterRelationWithHasMany(): HasMany
     {
         return $this->hasMany(User::class)->active();
     }
 
+    /** @phpstan-return Builder<User> */
     public function testScopeAfterQueryBuilderStaticCall(): Builder
     {
         return User::where('name', 'bar')->where('name', 'bar')->where('name', 'bar')->where('name', 'bar')->where('name', 'bar')->where('name', 'bar')->where('name', 'bar')->where('name', 'bar')->where('name', 'bar')->active();
     }
 
+    /** @phpstan-return Builder<User> */
     public function testScopeAfterQueryBuilderVariableCall(): Builder
     {
         $this->user = new User;
@@ -44,6 +48,7 @@ class Scopes extends Model
         return $this->user->where('name', 'bar')->active()->first();
     }
 
+    /** @phpstan-return Builder<Scopes> */
     public function testVoidScopeStillReturnsBuilder(): Builder
     {
         return $this->withVoidReturn();
@@ -54,6 +59,7 @@ class Scopes extends Model
         return $this->withVoidReturn()->first();
     }
 
+    /** @phpstan-param Builder<Scopes> $query */
     public function scopeWithVoidReturn(Builder $query): void
     {
         $query->where('whyuse', 'void');
